Move UI classes to .ui package
